Turtle Game
I studied this code and used my knowledge aslo manipulating some code lines I saw. I was able to come up with a turtle that I could move around with the keys W, A , S, D for its movement up and down.
This obviously is not the best code but that I was able to come up with something working increases my desire to learn coding the more rather tham killing my interest in it. I need more os encouragement as I write.
I would be glad to get more encouragement.
Thank you